+++ changelog.md
@@ -1,5 +1,8 @@
 # changelog
 
+### Wednesday, 25/05/2016
+- The class table generators have the proper ability score improvement progression.
+
 ### Tuesday, 24/05/2016 - v2.0.4
 - Fixed extra wide monster stat blocks sometimes only being one column
 - The class table generators now follow the proper progression from the PHB (thakns u/IrishBandit)